Add reference data for child elements that can exist inside of text and shield symbolizers.
Properties are defined for child elements in the same format as symbolizer properties.
Each element contains data about one or more psuedo-elements that can be used to create the
element with carto css. An array is defined inside of each symbolizer and element that can
potentially contain child elements to specifiy which types of elements they support.